% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezier.R
\name{bezier}
\alias{bezier}
\alias{qbezier}
\alias{cbezier}
\title{Bezier Curves}
\usage{
qbezier(x, y, t)

cbezier(x, y, t)
}
\arguments{
\item{x}{a numeric vector of x-coordinates for the control points}

\item{y}{a numeric vector of y-coordinates for the control points}

\item{t}{a numeric vector of parameter values at which to evaluate the curve}
}
\value{
A list with components \code{x} and \code{y}, each a numeric vector
of the same length as \code{t}.
}
\description{
Compute quadratic and cubic Bezier curves for the given control points.
}
\details{
\code{qbezier} computes the quadratic Bezier curve through three control
points and \code{cbezier} computes the cubic Bezier curve through four
control points. The curve is evaluated at all values in the vector
\code{t}, which typically ranges from 0 to 1.
}
\examples{
x <- c(1, 2, 3)
y <- c(2, 3, 5)
f <- qbezier(x, y, seq(0, 1, 1/100))

x <- c(-1, 1, 0, -2)
y <- c(-2, 2, -1, -1)
f <- cbezier(x, y, seq(0, 1, 1/100))

}
\seealso{
Other interp:
\code{\link[=bilinear]{bilinear()}},
\code{\link[=cubicspline]{cubicspline()}},
\code{\link[=linterp]{linterp()}},
\code{\link[=nn]{nn()}},
\code{\link[=polyinterp]{polyinterp()}},
\code{\link[=pwiselinterp]{pwiselinterp()}}
}
\concept{interp}
