% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilinear.R
\name{bilinear}
\alias{bilinear}
\title{Bilinear Interpolation}
\usage{
bilinear(x, y, z, newx, newy)
}
\arguments{
\item{x}{a numeric vector of length 2 giving \eqn{x_1} and \eqn{x_2}}

\item{y}{a numeric vector of length 2 giving \eqn{y_1} and \eqn{y_2}}

\item{z}{a 2-by-2 numeric matrix of z-values at the four corners}

\item{newx}{a numeric vector of x-coordinates at which to interpolate}

\item{newy}{a numeric vector of y-coordinates at which to interpolate}
}
\value{
A numeric vector of interpolated z-values.
}
\description{
Compute a bilinear interpolation bounded by four corner points.
}
\details{
\code{bilinear} performs bilinear interpolation within the rectangle
defined by the corners \eqn{(x_1, y_1)} and \eqn{(x_2, y_2)}.
First the function interpolates along the x-axis, then along the
y-axis.
}
\examples{
x <- c(2, 4)
y <- c(4, 7)
z <- matrix(c(81, 84, 85, 89), nrow = 2)
newx <- c(2.5, 3, 3.5)
newy <- c(5, 5.5, 6)
bilinear(x, y, z, newx, newy)

}
\seealso{
Other interp:
\code{\link{bezier}},
\code{\link[=cubicspline]{cubicspline()}},
\code{\link[=linterp]{linterp()}},
\code{\link[=nn]{nn()}},
\code{\link[=polyinterp]{polyinterp()}},
\code{\link[=pwiselinterp]{pwiselinterp()}}

Other algebra:
\code{\link[=cubicspline]{cubicspline()}},
\code{\link{division}},
\code{\link[=fibonacci]{fibonacci()}},
\code{\link[=isPrime]{isPrime()}},
\code{\link[=linterp]{linterp()}},
\code{\link[=nthroot]{nthroot()}},
\code{\link[=polyinterp]{polyinterp()}},
\code{\link{polynomial-evaluation}},
\code{\link[=pwiselinterp]{pwiselinterp()}},
\code{\link[=quadratic]{quadratic()}}
}
\concept{algebra}
\concept{interp}
