% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvp.R
\name{bvp}
\alias{bvp}
\alias{bvpexample}
\alias{bvpexample10}
\title{Boundary Value Problem Example}
\usage{
bvpexample(x)

bvpexample10(x)
}
\arguments{
\item{x}{a finite numeric scalar giving the proposed initial slope}
}
\value{
The residual \eqn{y(1) - 1} for the given trial slope.
}
\description{
Example functions for solving boundary value problems via the
shooting method.
}
\details{
\code{bvpexample} solves a boundary value problem using the shooting
method with \code{\link[=eulersys]{eulersys()}}. The ODE system is \eqn{y'' = y^2 - 2}
with boundary conditions \eqn{y(0) = 1} and \eqn{y(1) = 1}. The
parameter \code{x} is the trial value for \eqn{y'(0)}. The return value
is the residual at \eqn{x = 1}, suitable for use with a root-finding
method such as \code{\link[=bisection]{bisection()}} or \code{\link[=secant]{secant()}}.

\code{bvpexample10} is the same problem solved with 10 steps instead of
1000, for faster but less accurate evaluation.
}
\examples{
bvpexample(-2)
bvpexample(-1)
bvpexample(0)
bvpexample(1)
bvpexample(2)
## (bvp.b <- bisection(bvpexample, 0, 1))
## (bvp.s <- secant(bvpexample, 0))

}
