% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cholesky.R
\name{choleskymatrix}
\alias{choleskymatrix}
\title{Cholesky decomposition}
\usage{
choleskymatrix(m)
}
\arguments{
\item{m}{a square numeric matrix (must be symmetric positive-definite)}
}
\value{
The upper-triangular matrix \code{L}.
}
\description{
Decompose a symmetric positive-definite matrix into its Cholesky
factorization.
}
\details{
\code{choleskymatrix} decomposes the matrix \code{m} into the Cholesky
decomposition, such that \eqn{m = L^T \times L}.
}
\examples{
(A <- matrix(c(5, 1, 2, 1, 9, 3, 2, 3, 7), 3))
(L <- choleskymatrix(A))
t(L) \%*\% L

}
\seealso{
Other linear:
\code{\link[=detmatrix]{detmatrix()}},
\code{\link[=gdls]{gdls()}},
\code{\link[=invmatrix]{invmatrix()}},
\code{\link{iterativematrix}},
\code{\link[=lumatrix]{lumatrix()}},
\code{\link[=refmatrix]{refmatrix()}},
\code{\link{rowops}},
\code{\link[=tridiagmatrix]{tridiagmatrix()}},
\code{\link[=vecnorm]{vecnorm()}}
}
\concept{linear}
