% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubicspline.R
\name{cubicspline}
\alias{cubicspline}
\title{Natural Cubic Spline Interpolation}
\usage{
cubicspline(x, y)
}
\arguments{
\item{x}{a numeric vector of x-coordinates}

\item{y}{a numeric vector of y-coordinates}
}
\value{
A list with components \code{a}, \code{b}, \code{c}, and \code{d}, each of
length \code{length(x) - 1}, giving the cubic polynomial coefficients.
}
\description{
Find a natural cubic spline that interpolates the data points.
}
\details{
\code{cubicspline} computes a natural cubic spline interpolant for the
given data points. The vectors \code{x} and \code{y} must have the same
length. The function returns a list of four coefficient vectors
representing the piecewise cubic polynomials.
}
\examples{
x <- c(1, 2, 3)
y <- c(2, 3, 5)
f <- cubicspline(x, y)

x <- c(-1, 1, 0, -2)
y <- c(-2, 2, -1, -1)
f <- cubicspline(x, y)

}
\seealso{
Other interp:
\code{\link{bezier}},
\code{\link[=bilinear]{bilinear()}},
\code{\link[=linterp]{linterp()}},
\code{\link[=nn]{nn()}},
\code{\link[=polyinterp]{polyinterp()}},
\code{\link[=pwiselinterp]{pwiselinterp()}}

Other algebra:
\code{\link[=bilinear]{bilinear()}},
\code{\link{division}},
\code{\link[=fibonacci]{fibonacci()}},
\code{\link[=isPrime]{isPrime()}},
\code{\link[=linterp]{linterp()}},
\code{\link[=nthroot]{nthroot()}},
\code{\link[=polyinterp]{polyinterp()}},
\code{\link{polynomial-evaluation}},
\code{\link[=pwiselinterp]{pwiselinterp()}},
\code{\link[=quadratic]{quadratic()}}
}
\concept{algebra}
\concept{interp}
