% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detmatrix.R
\name{detmatrix}
\alias{detmatrix}
\title{Calculate the determinant of a matrix}
\usage{
detmatrix(m)
}
\arguments{
\item{m}{a square numeric matrix}
}
\value{
A numeric scalar giving the determinant.
}
\description{
Calculate the determinant of a square matrix.
}
\details{
\code{detmatrix} calculates the determinant of the matrix given by
reducing it to row echelon form and taking the product of the
diagonal.
}
\examples{
A <- matrix(c(1, 2, -7, -1, -1, 1, 2, 1, 5), 3)
detmatrix(A)

}
\seealso{
Other linear:
\code{\link[=choleskymatrix]{choleskymatrix()}},
\code{\link[=gdls]{gdls()}},
\code{\link[=invmatrix]{invmatrix()}},
\code{\link{iterativematrix}},
\code{\link[=lumatrix]{lumatrix()}},
\code{\link[=refmatrix]{refmatrix()}},
\code{\link{rowops}},
\code{\link[=tridiagmatrix]{tridiagmatrix()}},
\code{\link[=vecnorm]{vecnorm()}}
}
\concept{linear}
