% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naivediv.R
\name{division}
\alias{division}
\alias{naivediv}
\alias{longdiv}
\title{Integer division algorithms}
\usage{
naivediv(m, n)

longdiv(m, n)
}
\arguments{
\item{m}{the dividend (a nonnegative integer)}

\item{n}{the divisor (a positive integer)}
}
\value{
A list with components \code{quotient} and \code{remainder}.
}
\description{
Compute the quotient and remainder of integer division using
elementary algorithms.
}
\details{
\code{naivediv} divides \code{m} by \code{n} using repeated subtraction.
\code{longdiv} uses the binary long-division algorithm, processing
bits from most significant to least significant.

Both functions return the unique quotient \code{q} and remainder \code{r}
satisfying \code{m = q * n + r} with \verb{0 <= r < n}.
}
\examples{
naivediv(17, 5)
longdiv(17, 5)

}
\seealso{
Other algebra:
\code{\link[=bilinear]{bilinear()}},
\code{\link[=cubicspline]{cubicspline()}},
\code{\link[=fibonacci]{fibonacci()}},
\code{\link[=isPrime]{isPrime()}},
\code{\link[=linterp]{linterp()}},
\code{\link[=nthroot]{nthroot()}},
\code{\link[=polyinterp]{polyinterp()}},
\code{\link{polynomial-evaluation}},
\code{\link[=pwiselinterp]{pwiselinterp()}},
\code{\link[=quadratic]{quadratic()}}
}
\concept{algebra}
