% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdls.R
\name{gdls}
\alias{gdls}
\title{Least squares with gradient descent}
\usage{
gdls(A, b, alpha = 0.05, tol = 1e-06, m = 1e+05)
}
\arguments{
\item{A}{a matrix representing the design matrix of a linear system}

\item{b}{a numeric vector representing the response}

\item{alpha}{the learning rate (positive numeric scalar)}

\item{tol}{the convergence tolerance (positive numeric scalar)}

\item{m}{the maximum number of iterations (positive integer)}
}
\value{
The coefficient vector.
}
\description{
Solve least squares with gradient descent.
}
\details{
\code{gdls} solves a linear system using gradient descent to minimize
the sum of squared residuals.
}
\examples{
head(b <- iris$Sepal.Length)
head(A <- matrix(cbind(1, iris$Sepal.Width, iris$Petal.Length, iris$Petal.Width), ncol = 4))
gdls(A, b, alpha = 0.05, m = 10000)

}
\seealso{
Other linear:
\code{\link[=choleskymatrix]{choleskymatrix()}},
\code{\link[=detmatrix]{detmatrix()}},
\code{\link[=invmatrix]{invmatrix()}},
\code{\link{iterativematrix}},
\code{\link[=lumatrix]{lumatrix()}},
\code{\link[=refmatrix]{refmatrix()}},
\code{\link{rowops}},
\code{\link[=tridiagmatrix]{tridiagmatrix()}},
\code{\link[=vecnorm]{vecnorm()}}
}
\concept{linear}
