% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goldsect.R
\name{goldsect}
\alias{goldsect}
\alias{goldsectmin}
\alias{goldsectmax}
\title{Golden Section Search}
\usage{
goldsectmin(f, a, b, tol = 0.001, m = 100)

goldsectmax(f, a, b, tol = 0.001, m = 100)
}
\arguments{
\item{f}{function to optimize}

\item{a}{lower bound of the search region (finite numeric scalar)}

\item{b}{upper bound of the search region (finite numeric scalar)}

\item{tol}{error tolerance (positive numeric scalar, default 1e-3)}

\item{m}{maximum number of iterations (positive integer, default 100)}
}
\value{
The \code{x} value of the extremum found.
}
\description{
Use golden section search to find local extrema.
}
\details{
The golden section search method functions by repeatedly dividing
the interval between \code{a} and \code{b} using the golden ratio and will
return when the interval between them is less than \code{tol}, the error
tolerance. However, this implementation also stops if after \code{m}
iterations.

\code{goldsectmin} searches for a local minimum; \code{goldsectmax} searches
for a local maximum.
}
\examples{
f <- function(x) { x^2 - 3 * x + 3 }
goldsectmin(f, 0, 5)

}
\seealso{
Other optimz:
\code{\link{gradient}},
\code{\link[=hillclimbing]{hillclimbing()}},
\code{\link[=sa]{sa()}}
}
\concept{optimz}
