% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{heat}
\alias{heat}
\title{Heat Equation via Forward-Time Central-Space}
\usage{
heat(u, alpha, xdelta, tdelta, n)
}
\arguments{
\item{u}{a numeric vector of initial values}

\item{alpha}{a positive numeric scalar giving the thermal diffusivity
coefficient}

\item{xdelta}{a positive numeric scalar giving the spatial step size}

\item{tdelta}{a positive numeric scalar giving the time step size}

\item{n}{a positive integer giving the number of time steps}
}
\value{
A matrix with \code{n + 1} rows, where each row is the solution
vector \code{u} at a given time step.
}
\description{
Solve the one-dimensional heat equation using the forward-time
central-space (FTCS) finite difference method.
}
\details{
\code{heat} solves the heat equation using the forward-time central-space
method in one dimension. Periodic boundary conditions are applied by
copying the last element to the first at each time step.
}
\examples{
alpha <- 1
x0 <- 0
xdelta <- .05
x <- seq(x0, 1, xdelta)
u <- sin(x^4 * pi)
tdelta <- .001
n <- 25
z <- heat(u, alpha, xdelta, tdelta, n)

}
