% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillclimbing.R
\name{hillclimbing}
\alias{hillclimbing}
\title{Hill Climbing}
\usage{
hillclimbing(f, x, h = 1, m = 1000)
}
\arguments{
\item{f}{function to minimize}

\item{x}{an initial estimate of the minimum (numeric vector)}

\item{h}{step size for random perturbation (positive numeric scalar,
default 1)}

\item{m}{maximum number of iterations (positive integer, default 1000)}
}
\value{
The \code{x} value of the minimum found.
}
\description{
Use hill climbing to find a local minimum.
}
\details{
Hill climbing is a stochastic optimization method that iteratively
perturbs a randomly chosen component of \code{x} by drawing from a normal
distribution with standard deviation \code{h}. If the perturbation
improves the objective value, the new point is accepted. The process
repeats for \code{m} iterations.
}
\examples{
f <- function(x) {
    (x[1]^2 + x[2] - 11)^2 + (x[1] + x[2]^2 - 7)^2
}
hillclimbing(f, c(0,0))
hillclimbing(f, c(-1,-1))
hillclimbing(f, c(10,10))

}
\seealso{
Other optimz:
\code{\link{goldsect}},
\code{\link{gradient}},
\code{\link[=sa]{sa()}}
}
\concept{optimz}
