% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invmatrix.R
\name{invmatrix}
\alias{invmatrix}
\title{Invert a matrix}
\usage{
invmatrix(m)
}
\arguments{
\item{m}{a square numeric matrix}
}
\value{
The inverted matrix.
}
\description{
Invert a matrix using Gaussian elimination.
}
\details{
\code{invmatrix} inverts the given matrix using Gaussian elimination
and returns the result.
}
\examples{
A <- matrix(c(1, 2, -7, -1, -1, 1, 2, 1, 5), 3)
invmatrix(A)

}
\seealso{
Other linear:
\code{\link[=choleskymatrix]{choleskymatrix()}},
\code{\link[=detmatrix]{detmatrix()}},
\code{\link[=gdls]{gdls()}},
\code{\link{iterativematrix}},
\code{\link[=lumatrix]{lumatrix()}},
\code{\link[=refmatrix]{refmatrix()}},
\code{\link{rowops}},
\code{\link[=tridiagmatrix]{tridiagmatrix()}},
\code{\link[=vecnorm]{vecnorm()}}
}
\concept{linear}
