% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isPrime.R
\name{isPrime}
\alias{isPrime}
\title{Test for primality}
\usage{
isPrime(n)
}
\arguments{
\item{n}{a positive integer to test}
}
\value{
\code{TRUE} if \code{n} is prime, \code{FALSE} otherwise.
}
\description{
Test whether a positive integer is prime by trial division.
}
\details{
\code{isPrime} tests \code{n} for primality by attempting division by every
integer from 2 up to \code{floor(sqrt(n))}. If any divisor produces a
zero remainder, the number is composite.

This is the simplest correct primality test and is efficient only
for small values of \code{n}.
}
\examples{
isPrime(2)
isPrime(37)
isPrime(100)

}
\seealso{
Other algebra:
\code{\link[=bilinear]{bilinear()}},
\code{\link[=cubicspline]{cubicspline()}},
\code{\link{division}},
\code{\link[=fibonacci]{fibonacci()}},
\code{\link[=linterp]{linterp()}},
\code{\link[=nthroot]{nthroot()}},
\code{\link[=polyinterp]{polyinterp()}},
\code{\link{polynomial-evaluation}},
\code{\link[=pwiselinterp]{pwiselinterp()}},
\code{\link[=quadratic]{quadratic()}}
}
\concept{algebra}
