% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterativematrix.R
\name{iterativematrix}
\alias{iterativematrix}
\alias{jacobi}
\alias{gaussseidel}
\alias{cgmmatrix}
\title{Solve a matrix using iterative methods}
\usage{
jacobi(A, b, tol = 1e-06, maxiter = 100)

gaussseidel(A, b, tol = 1e-06, maxiter = 100)

cgmmatrix(A, b, tol = 1e-06, maxiter = 100)
}
\arguments{
\item{A}{a square numeric matrix representing the coefficients of a
linear system}

\item{b}{a numeric vector representing the right-hand side of the
linear system}

\item{tol}{a numeric tolerance for convergence}

\item{maxiter}{the maximum number of iterations}
}
\value{
The solution vector.
}
\description{
Solve a linear system using iterative methods.
}
\details{
\code{jacobi} finds the solution using Jacobi iteration.
Jacobi iteration depends on the matrix being diagonally dominant.
The tolerance is measured by the norm of the solution vector.

\code{gaussseidel} finds the solution using Gauss-Seidel iteration.
Gauss-Seidel iteration depends on the matrix being either
diagonally dominant or symmetric and positive definite.

\code{cgmmatrix} finds the solution using the conjugate gradient
method. The conjugate gradient method depends on the matrix being
symmetric and positive definite.
}
\examples{
A <- matrix(c(5, 2, 1, 2, 7, 3, 3, 4, 8), 3)
b <- c(40, 39, 55)
jacobi(A, b)

}
\seealso{
Other linear:
\code{\link[=choleskymatrix]{choleskymatrix()}},
\code{\link[=detmatrix]{detmatrix()}},
\code{\link[=gdls]{gdls()}},
\code{\link[=invmatrix]{invmatrix()}},
\code{\link[=lumatrix]{lumatrix()}},
\code{\link[=refmatrix]{refmatrix()}},
\code{\link{rowops}},
\code{\link[=tridiagmatrix]{tridiagmatrix()}},
\code{\link[=vecnorm]{vecnorm()}}
}
\concept{linear}
