% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linterp.R
\name{linterp}
\alias{linterp}
\title{Linear Interpolation}
\usage{
linterp(x1, y1, x2, y2)
}
\arguments{
\item{x1}{a finite numeric scalar giving the x-coordinate of the first point}

\item{y1}{a finite numeric scalar giving the y-coordinate of the first point}

\item{x2}{a finite numeric scalar giving the x-coordinate of the second point}

\item{y2}{a finite numeric scalar giving the y-coordinate of the second point}
}
\value{
A numeric vector of length 2 giving the intercept and slope.
}
\description{
Find the linear function passing through two points.
}
\details{
\code{linterp} computes the coefficients of the unique linear function
passing through the two points \eqn{(x_1, y_1)} and \eqn{(x_2, y_2)}.
The coefficients are returned in a form suitable for \code{\link[=horner]{horner()}}.
}
\examples{
f <- linterp(3, 2, 7, -2)

}
\seealso{
Other interp:
\code{\link{bezier}},
\code{\link[=bilinear]{bilinear()}},
\code{\link[=cubicspline]{cubicspline()}},
\code{\link[=nn]{nn()}},
\code{\link[=polyinterp]{polyinterp()}},
\code{\link[=pwiselinterp]{pwiselinterp()}}

Other algebra:
\code{\link[=bilinear]{bilinear()}},
\code{\link[=cubicspline]{cubicspline()}},
\code{\link{division}},
\code{\link[=fibonacci]{fibonacci()}},
\code{\link[=isPrime]{isPrime()}},
\code{\link[=nthroot]{nthroot()}},
\code{\link[=polyinterp]{polyinterp()}},
\code{\link{polynomial-evaluation}},
\code{\link[=pwiselinterp]{pwiselinterp()}},
\code{\link[=quadratic]{quadratic()}}
}
\concept{algebra}
\concept{interp}
