% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumatrix.R
\name{lumatrix}
\alias{lumatrix}
\title{LU decomposition}
\usage{
lumatrix(m)
}
\arguments{
\item{m}{a square numeric matrix}
}
\value{
A list with matrices \code{P}, \code{L}, and \code{U} representing the
LU decomposition.
}
\description{
Decompose a square matrix into lower- and upper-triangular matrices.
}
\details{
\code{lumatrix} decomposes the matrix \code{m} into the LU decomposition,
such that \eqn{P \cdot m = L \times U}.
}
\examples{
A <- matrix(c(1, 2, -7, -1, -1, 1, 2, 1, 5), 3)
lumatrix(A)

}
\seealso{
Other linear:
\code{\link[=choleskymatrix]{choleskymatrix()}},
\code{\link[=detmatrix]{detmatrix()}},
\code{\link[=gdls]{gdls()}},
\code{\link[=invmatrix]{invmatrix()}},
\code{\link{iterativematrix}},
\code{\link[=refmatrix]{refmatrix()}},
\code{\link{rowops}},
\code{\link[=tridiagmatrix]{tridiagmatrix()}},
\code{\link[=vecnorm]{vecnorm()}}
}
\concept{linear}
