% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{nn}
\alias{nn}
\title{Nearest Neighbor Interpolation}
\usage{
nn(p, y, q)
}
\arguments{
\item{p}{a numeric matrix of variable values, where each row is a data point}

\item{y}{a numeric vector of values, each entry corresponding to one row
in \code{p}}

\item{q}{a numeric matrix of query points, where each row is a point to
interpolate and the number of columns must match \code{p}}
}
\value{
The interpolated value from \code{y} for the nearest neighbor of \code{q}.
}
\description{
Find the nearest neighbor for a set of data points.
}
\details{
\code{nn} finds the n-dimensional nearest neighbor among the rows of \code{p}
for the query point \code{q}, using Euclidean distance, and returns the
corresponding value from \code{y}.
}
\examples{
p <- matrix(floor(runif(100, 0, 9)), 20)
y <- floor(runif(20, 0, 9))
q <- matrix(floor(runif(5, 0, 9)), 1)
nn(p, y, q)

}
\seealso{
Other interp:
\code{\link{bezier}},
\code{\link[=bilinear]{bilinear()}},
\code{\link[=cubicspline]{cubicspline()}},
\code{\link[=linterp]{linterp()}},
\code{\link[=polyinterp]{polyinterp()}},
\code{\link[=pwiselinterp]{pwiselinterp()}}
}
\concept{interp}
