% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyinterp.R
\name{polyinterp}
\alias{polyinterp}
\title{Polynomial Interpolation}
\usage{
polyinterp(x, y)
}
\arguments{
\item{x}{a numeric vector of x-coordinates}

\item{y}{a numeric vector of y-coordinates}
}
\value{
A numeric vector of polynomial coefficients in ascending
order of degree, suitable for \code{\link[=horner]{horner()}}.
}
\description{
Find the polynomial that interpolates the given points.
}
\details{
\code{polyinterp} constructs the Vandermonde matrix for the given
x-coordinates and solves for the polynomial coefficients that
interpolate the corresponding y-values. The vectors \code{x} and \code{y}
must have the same length.
}
\examples{
x <- c(1, 2, 3)
y <- x^2 + 5 * x - 3
f <- polyinterp(x, y)

}
\seealso{
Other interp:
\code{\link{bezier}},
\code{\link[=bilinear]{bilinear()}},
\code{\link[=cubicspline]{cubicspline()}},
\code{\link[=linterp]{linterp()}},
\code{\link[=nn]{nn()}},
\code{\link[=pwiselinterp]{pwiselinterp()}}

Other algebra:
\code{\link[=bilinear]{bilinear()}},
\code{\link[=cubicspline]{cubicspline()}},
\code{\link{division}},
\code{\link[=fibonacci]{fibonacci()}},
\code{\link[=isPrime]{isPrime()}},
\code{\link[=linterp]{linterp()}},
\code{\link[=nthroot]{nthroot()}},
\code{\link{polynomial-evaluation}},
\code{\link[=pwiselinterp]{pwiselinterp()}},
\code{\link[=quadratic]{quadratic()}}
}
\concept{algebra}
\concept{interp}
