% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwiselinterp.R
\name{pwiselinterp}
\alias{pwiselinterp}
\title{Piecewise Linear Interpolation}
\usage{
pwiselinterp(x, y)
}
\arguments{
\item{x}{a numeric vector of x-coordinates}

\item{y}{a numeric vector of y-coordinates}
}
\value{
A list with components \code{m} (slopes) and \code{b} (intercepts),
each of length \code{length(x) - 1}.
}
\description{
Find a piecewise linear function that interpolates the data points.
}
\details{
\code{pwiselinterp} sorts the data by x-coordinate, then finds the unique
line interpolating each consecutive pair of points. The result is a
list with two vectors of slope and intercept coefficients from the
slope-intercept form \eqn{y = mx + b}.
}
\examples{
x <- c(5, 0, 3)
y <- c(4, 0, 3)
f <- pwiselinterp(x, y)

}
\seealso{
Other interp:
\code{\link{bezier}},
\code{\link[=bilinear]{bilinear()}},
\code{\link[=cubicspline]{cubicspline()}},
\code{\link[=linterp]{linterp()}},
\code{\link[=nn]{nn()}},
\code{\link[=polyinterp]{polyinterp()}}

Other algebra:
\code{\link[=bilinear]{bilinear()}},
\code{\link[=cubicspline]{cubicspline()}},
\code{\link{division}},
\code{\link[=fibonacci]{fibonacci()}},
\code{\link[=isPrime]{isPrime()}},
\code{\link[=linterp]{linterp()}},
\code{\link[=nthroot]{nthroot()}},
\code{\link[=polyinterp]{polyinterp()}},
\code{\link{polynomial-evaluation}},
\code{\link[=quadratic]{quadratic()}}
}
\concept{algebra}
\concept{interp}
