% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refmatrix.R
\name{refmatrix}
\alias{refmatrix}
\alias{rrefmatrix}
\alias{solvematrix}
\title{Matrix to row echelon form}
\usage{
refmatrix(m)

rrefmatrix(m)

solvematrix(A, b)
}
\arguments{
\item{m}{a matrix}

\item{A}{a matrix representing the coefficients of a linear
system in \code{solvematrix}}

\item{b}{a numeric vector representing the right-hand side of the
linear system in \code{solvematrix}}
}
\value{
The modified matrix, or for \code{solvematrix} the solution vector.
}
\description{
Transform a matrix to row echelon form.
}
\details{
\code{refmatrix} reduces a matrix to row echelon form. This is not a
reduced row echelon form, though that can be easily calculated from
the diagonal. This function works on non-square matrices.

\code{rrefmatrix} returns the reduced row echelon matrix.

\code{solvematrix} solves a linear system using \code{rrefmatrix}.
}
\examples{
A <- matrix(c(1, 2, -7, -1, -1, 1, 2, 1, 5), 3)
refmatrix(A)

}
\seealso{
Other linear:
\code{\link[=choleskymatrix]{choleskymatrix()}},
\code{\link[=detmatrix]{detmatrix()}},
\code{\link[=gdls]{gdls()}},
\code{\link[=invmatrix]{invmatrix()}},
\code{\link{iterativematrix}},
\code{\link[=lumatrix]{lumatrix()}},
\code{\link{rowops}},
\code{\link[=tridiagmatrix]{tridiagmatrix()}},
\code{\link[=vecnorm]{vecnorm()}}
}
\concept{linear}
