% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resizeImage.R
\name{resizeImage}
\alias{resizeImage}
\alias{resizeImageNN}
\alias{resizeImageBL}
\title{Image Resizing}
\usage{
resizeImageNN(imx, width, height)

resizeImageBL(imx, width, height)
}
\arguments{
\item{imx}{a 3-dimensional numeric array containing image data}

\item{width}{a positive integer giving the new width}

\item{height}{a positive integer giving the new height}
}
\value{
A 3-dimensional array containing the resized image.
}
\description{
Resize images using nearest neighbor or bilinear interpolation.
}
\details{
\code{resizeImageNN} uses the nearest neighbor method to resize the image.
\code{resizeImageBL} uses bilinear interpolation to resize the image.
}
\concept{interpolation}
