% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revolution-solid.R
\name{revolution-solid}
\alias{revolution-solid}
\alias{shellmethod}
\alias{discmethod}
\title{Volumes of Solids of Revolution}
\usage{
shellmethod(f, a, b)

discmethod(f, a, b)
}
\arguments{
\item{f}{function defining the curve to revolve}

\item{a}{lower bound of the solid (finite numeric scalar)}

\item{b}{upper bound of the solid (finite numeric scalar)}
}
\value{
A numeric scalar giving the volume of the solid.
}
\description{
Find the volume of a solid of revolution using the shell or disc
method.
}
\details{
\code{shellmethod} computes the volume of the solid obtained by revolving
the curve \code{f(x)} about the \emph{y}-axis on \verb{[a, b]} using the
cylindrical-shell formula \code{2 * pi * integral(x * f(x))}.

\code{discmethod} computes the volume of the solid obtained by revolving
\code{f(x)} about the \emph{x}-axis on \verb{[a, b]} using the disc formula
\code{integral(pi * f(x)^2)}.
}
\examples{
f <- function(x) { x^2 }
shellmethod(f, 1, 2)
discmethod(f, 1, 2)

}
\seealso{
Other integration:
\code{\link[=adaptint]{adaptint()}},
\code{\link[=gaussint]{gaussint()}},
\code{\link[=giniquintile]{giniquintile()}},
\code{\link[=mcint]{mcint()}},
\code{\link[=midpt]{midpt()}},
\code{\link[=romberg]{romberg()}},
\code{\link[=simp]{simp()}},
\code{\link[=simp38]{simp38()}},
\code{\link[=trap]{trap()}}
}
\concept{integration}
