% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowops.R
\name{rowops}
\alias{rowops}
\alias{swaprows}
\alias{replacerow}
\alias{scalerow}
\title{Elementary row operations}
\usage{
swaprows(m, row1, row2)

replacerow(m, row1, row2, k)

scalerow(m, row, k)
}
\arguments{
\item{m}{a matrix}

\item{row1}{a source row index}

\item{row2}{a destination row index}

\item{k}{a numeric scaling factor}

\item{row}{a row index to modify}
}
\value{
The modified matrix.
}
\description{
Elementary row operations for a matrix. These do not require a
square matrix and will work on any matrix. They use R's internal
row addressing to function.
}
\details{
\code{replacerow} replaces one row with the sum of itself and the
multiple of another row. \code{swaprows} swaps two rows in the
matrix. \code{scalerow} scales all entries in a row by a constant.
}
\examples{
n <- 5
A <- matrix(sample.int(10, n^2, TRUE) - 1, n)
A <- swaprows(A, 2, 4)
A <- replacerow(A, 1, 3, 2)
A <- scalerow(A, 5, 10)
}
\seealso{
Other linear:
\code{\link[=choleskymatrix]{choleskymatrix()}},
\code{\link[=detmatrix]{detmatrix()}},
\code{\link[=gdls]{gdls()}},
\code{\link[=invmatrix]{invmatrix()}},
\code{\link{iterativematrix}},
\code{\link[=lumatrix]{lumatrix()}},
\code{\link[=refmatrix]{refmatrix()}},
\code{\link[=tridiagmatrix]{tridiagmatrix()}},
\code{\link[=vecnorm]{vecnorm()}}
}
\concept{linear}
