% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{sa}
\alias{sa}
\alias{tspsa}
\title{Simulated Annealing}
\usage{
sa(f, x, temp = 10000, rate = 1e-04)

tspsa(x, temp = 100, rate = 1e-04)
}
\arguments{
\item{f}{function to minimize}

\item{x}{an initial estimate of the minimum (numeric vector for \code{sa},
numeric matrix of point coordinates for \code{tspsa})}

\item{temp}{initial temperature (positive numeric scalar)}

\item{rate}{cooling rate (positive numeric scalar between 0 and 1)}
}
\value{
For \code{sa}, the \code{x} value of the minimum found. For \code{tspsa},
a list with components \code{order} (the best tour) and \code{distance}
(the total tour distance).
}
\description{
Use simulated annealing to find the global minimum.
}
\details{
Simulated annealing finds a global minimum by mimicking the
metallurgical process of annealing. At each step a random
perturbation is proposed and accepted if it improves the objective
or with probability depending on the current temperature.

\code{sa} performs general-purpose simulated annealing on a numeric
vector. \code{tspsa} solves the travelling salesman problem on a matrix
of point coordinates, returning the best tour order and its total
distance.
}
\examples{
f <- function(x) { x^6 - 4 * x^5 - 7 * x^4 + 22 * x^3 + 24 * x^2 + 2}
sa(f, 0)

f <- function(x) { (x[1] - 1)^2 + (x[2] - 1)^2 }
sa(f, c(0, 0), 0.05)

}
\seealso{
Other optimz:
\code{\link{goldsect}},
\code{\link{gradient}},
\code{\link[=hillclimbing]{hillclimbing()}}
}
\concept{optimz}
