% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tridiagmatrix.R
\name{tridiagmatrix}
\alias{tridiagmatrix}
\title{Solve a tridiagonal matrix}
\usage{
tridiagmatrix(L, D, U, b)
}
\arguments{
\item{L}{numeric vector of entries below the main diagonal}

\item{D}{numeric vector of entries on the main diagonal}

\item{U}{numeric vector of entries above the main diagonal}

\item{b}{numeric vector of the right-hand side of the linear system}
}
\value{
The solution vector.
}
\description{
Use the tridiagonal matrix algorithm to solve a tridiagonal system.
}
\details{
\code{tridiagmatrix} uses the tridiagonal matrix algorithm (Thomas
algorithm) to solve a tridiagonal linear system.
}
\seealso{
Other linear:
\code{\link[=choleskymatrix]{choleskymatrix()}},
\code{\link[=detmatrix]{detmatrix()}},
\code{\link[=gdls]{gdls()}},
\code{\link[=invmatrix]{invmatrix()}},
\code{\link{iterativematrix}},
\code{\link[=lumatrix]{lumatrix()}},
\code{\link[=refmatrix]{refmatrix()}},
\code{\link{rowops}},
\code{\link[=vecnorm]{vecnorm()}}
}
\concept{linear}
