% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecnorm.R
\name{vecnorm}
\alias{vecnorm}
\title{Norm of a vector}
\usage{
vecnorm(b)
}
\arguments{
\item{b}{a numeric vector}
}
\value{
A numeric scalar giving the norm.
}
\description{
Compute the Euclidean (L2) norm of a numeric vector.
}
\details{
\code{vecnorm} computes the Euclidean norm of \code{b}, defined as
\eqn{\sqrt{\sum b_i^2}}.
}
\examples{
x <- c(1, 2, 3)
vecnorm(x)

}
\seealso{
Other linear:
\code{\link[=choleskymatrix]{choleskymatrix()}},
\code{\link[=detmatrix]{detmatrix()}},
\code{\link[=gdls]{gdls()}},
\code{\link[=invmatrix]{invmatrix()}},
\code{\link{iterativematrix}},
\code{\link[=lumatrix]{lumatrix()}},
\code{\link[=refmatrix]{refmatrix()}},
\code{\link{rowops}},
\code{\link[=tridiagmatrix]{tridiagmatrix()}}
}
\concept{linear}
