% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cholesky.R
\name{choleskymatrix}
\alias{choleskymatrix}
\title{Cholesky Decomposition}
\usage{
choleskymatrix(m)
}
\arguments{
\item{m}{a matrix}
}
\value{
the matrix L
}
\description{
Decompose a matrix into the Cholesky
}
\details{
\code{choleskymatrix} decomposes the matrix \code{m} into the LU
decomposition, such that m == L %*% L*.
}
\examples{
A <- matrix(c(5,1,2,1,9,3,2,3,7),3)
choleskymatrix(m)
}
\seealso{
Other linear: \code{\link{cgmmatrix}},
  \code{\link{gaussseidel}}, \code{\link{iterativematrix}},
  \code{\link{jacobi}}; \code{\link{detmatrix}};
  \code{\link{invmatrix}}; \code{\link{lumatrix}};
  \code{\link{refmatrix}}, \code{\link{rrefmatrix}},
  \code{\link{solvematrix}}; \code{\link{replacerow}},
  \code{\link{rowops}}, \code{\link{scalerow}},
  \code{\link{swaprows}}; \code{\link{tridiagmatrix}};
  \code{\link{vecnorm}}
}

