% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmna.R
\docType{package}
\name{cmna-package}
\alias{cmna}
\alias{cmna-package}
\title{Computational Methods for Numerical Analysis}
\description{
Provides the source and examples for \emph{Computational Methods for
Numerical Analysis with R}.
}
\details{
This package provides a suite of simple implementations of standard
methods from numerical analysis.  The collection is designed to
accompany \emph{Computational Methods for Numerical Analysis with R}
by James P. Howard, II.  Together, these functions provide methods to
support linear algebra, interpolation, integration, root finding,
optimization, and differential equations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://howardjp.github.io/cmna/}
  \item Report bugs at \url{https://github.com/howardjp/cmna/issues}
}

}
\author{
James P. Howard, II <jh@jameshoward.us>
}
