% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detmatrix.R
\name{detmatrix}
\alias{detmatrix}
\title{Calculate the determinant of the matrix}
\usage{
detmatrix(m)
}
\arguments{
\item{m}{a matrix}
}
\value{
the determinant
}
\description{
Calculate the determinant of the matrix
}
\details{
\code{detmatrix} calculates the determinant of the matrix given.
}
\examples{
A <- matrix(c(1, 2, -7, -1, -1, 1, 2, 1, 5), 3)
detmatrix(A)
}
\seealso{
Other linear: \code{\link{cgmmatrix}},
  \code{\link{gaussseidel}}, \code{\link{iterativematrix}},
  \code{\link{jacobi}}; \code{\link{choleskymatrix}};
  \code{\link{invmatrix}}; \code{\link{lumatrix}};
  \code{\link{refmatrix}}, \code{\link{rrefmatrix}},
  \code{\link{solvematrix}}; \code{\link{replacerow}},
  \code{\link{rowops}}, \code{\link{scalerow}},
  \code{\link{swaprows}}; \code{\link{tridiagmatrix}};
  \code{\link{vecnorm}}
}

