% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detmatrix.R
\name{detmatrix}
\alias{detmatrix}
\title{Calculate the determinant of the matrix}
\usage{
detmatrix(m)
}
\arguments{
\item{m}{a matrix}
}
\value{
the determinant
}
\description{
Calculate the determinant of the matrix
}
\details{
\code{detmatrix} calculates the determinant of the matrix given.
}
\examples{
A <- matrix(c(1, 2, -7, -1, -1, 1, 2, 1, 5), 3)
detmatrix(A)

}
\seealso{
Other linear: \code{\link{choleskymatrix}},
  \code{\link{gdls}}, \code{\link{invmatrix}},
  \code{\link{iterativematrix}}, \code{\link{lumatrix}},
  \code{\link{refmatrix}}, \code{\link{rowops}},
  \code{\link{tridiagmatrix}}, \code{\link{vecnorm}}
}

