% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fibonacci.R
\name{fibonacci}
\alias{fibonacci}
\title{Fibonacci numbers}
\usage{
fibonacci(n)
}
\arguments{
\item{n}{n}
}
\value{
the sequence element
}
\description{
Return the n-th Fibonacci number
}
\details{
This function is recursively implements the famous Fibonacci
sequence.  The function returns the \code{n}th member of the
sequence.
}
\examples{
fibonacci(10)
}
\seealso{
Other algebra: \code{\link{betterpoly}},
  \code{\link{horner}}, \code{\link{naivepoly}},
  \code{\link{rhorner}}; \code{\link{bilinear}};
  \code{\link{cubicspline}}; \code{\link{division}},
  \code{\link{longdiv}}, \code{\link{naivediv}};
  \code{\link{isPrime}}; \code{\link{linterp}};
  \code{\link{nthroot}}; \code{\link{polyinterp}};
  \code{\link{pwiselinterp}}; \code{\link{quadratic}},
  \code{\link{quadratic2}}
}

