% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giniquintile.R
\name{giniquintile}
\alias{giniquintile}
\title{Gini Coefficient from Quintile Data}
\usage{
giniquintile(L)
}
\arguments{
\item{L}{numeric vector of percentages at the 20th, 40th, 60th, and
80th percentiles}
}
\value{
The estimated Gini coefficient as a numeric scalar.
}
\description{
Calculate the Gini coefficient from quintile data.
}
\details{
Computes the Gini coefficient using Gerber's quintile rule, which
approximates the area between the Lorenz curve and the line of
equality from four cumulative income-share percentages.
}
\examples{
L <- c(4.3, 9.8, 15.4, 22.7)
giniquintile(L)

}
\references{
Leon Gerber, "A Quintile Rule for the Gini Coefficient",
\emph{Mathematics Magazine}, 80:2, April 2007.
}
\seealso{
Other integration:
\code{\link[=adaptint]{adaptint()}},
\code{\link[=gaussint]{gaussint()}},
\code{\link[=mcint]{mcint()}},
\code{\link[=midpt]{midpt()}},
\code{\link{revolution-solid}},
\code{\link[=romberg]{romberg()}},
\code{\link[=simp]{simp()}},
\code{\link[=simp38]{simp38()}},
\code{\link[=trap]{trap()}}

Other newton-cotes:
\code{\link[=adaptint]{adaptint()}},
\code{\link[=midpt]{midpt()}},
\code{\link[=romberg]{romberg()}},
\code{\link[=simp]{simp()}},
\code{\link[=simp38]{simp38()}},
\code{\link[=trap]{trap()}}
}
\concept{integration}
\concept{newton-cotes}
