% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/himmelblau.R
\name{himmelblau}
\alias{himmelblau}
\title{Himmelblau Function}
\usage{
himmelblau(x)
}
\arguments{
\item{x}{a vector of \code{x}-values}
}
\value{
the value of the function at \code{x}.
}
\description{
Generate the Himmelblau function
}
\details{
Generate the Himmelblau function
}

