% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invmatrix.R
\name{invmatrix}
\alias{invmatrix}
\title{Invert a matrix}
\usage{
invmatrix(m)
}
\arguments{
\item{m}{a matrix}
}
\value{
the inverted matrix
}
\description{
Invert the matrix using Gaussian elimination
}
\details{
\code{invmatrix} invertse the given matrix using Gaussian elimination
and returns the result.
}
\examples{
A <- matrix(c(1, 2, -7, -1, -1, 1, 2, 1, 5), 3)
refmatrix(A)

}
\seealso{
Other linear: \code{\link{choleskymatrix}},
  \code{\link{detmatrix}}, \code{\link{gdls}},
  \code{\link{iterativematrix}}, \code{\link{lumatrix}},
  \code{\link{refmatrix}}, \code{\link{rowops}},
  \code{\link{tridiagmatrix}}, \code{\link{vecnorm}}
}
\concept{linear}
