% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isPrime.R
\name{isPrime}
\alias{isPrime}
\title{Test for Primality}
\usage{
isPrime(n)
}
\arguments{
\item{n}{n}
}
\value{
boolean TRUE if \code{n} is prime, FALSE if not
}
\description{
Test the number given for primality.
}
\details{
This function tests \code{n} if it is prime through repeated division
attempts.  If a match is found, by finding a remainder of 0,
\code{FALSE} is returned.
}
\examples{
isPrime(37)
isPrime(89)
isPrime(100)

}
\seealso{
Other algebra: \code{\link{bilinear}},
  \code{\link{cubicspline}}, \code{\link{division}},
  \code{\link{fibonacci}}, \code{\link{horner}},
  \code{\link{linterp}}, \code{\link{nthroot}},
  \code{\link{polyinterp}}, \code{\link{pwiselinterp}},
  \code{\link{quadratic}}
}

