% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linterp.R
\name{linterp}
\alias{linterp}
\title{Linear interpolation}
\usage{
linterp(x1, y1, x2, y2)
}
\arguments{
\item{x1}{x value of the first point}

\item{y1}{y value of the first point}

\item{x2}{x value of the second point}

\item{y2}{y value of the second point}
}
\value{
a linear equation's coefficients
}
\description{
Finds a linear function between two points
}
\details{
\code{linterp} finds a linear function  between two points.
}
\examples{
f <- linterp(3, 2, 7, -2)
}
\seealso{
Other algebra: \code{\link{betterpoly}},
  \code{\link{horner}}, \code{\link{naivepoly}},
  \code{\link{rhorner}}; \code{\link{bilinear}};
  \code{\link{cubicspline}}; \code{\link{division}},
  \code{\link{longdiv}}, \code{\link{naivediv}};
  \code{\link{fibonacci}}; \code{\link{isPrime}};
  \code{\link{nthroot}}; \code{\link{polyinterp}};
  \code{\link{pwiselinterp}}; \code{\link{quadratic}},
  \code{\link{quadratic2}}

Other interp: \code{\link{bezier}}, \code{\link{cbezier}},
  \code{\link{qbezier}}; \code{\link{bilinear}};
  \code{\link{cubicspline}}; \code{\link{nn}};
  \code{\link{polyinterp}}; \code{\link{pwiselinterp}}
}

