% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simp38.R
\name{simp38}
\alias{simp38}
\title{Simpson's 3/8 rule}
\usage{
simp38(f, a, b, m = 100)
}
\arguments{
\item{f}{function to integrate}

\item{a}{the a-bound of integration}

\item{b}{the b-bound of integration}

\item{m}{the number of subintervals to calculate}
}
\value{
the value of the integral
}
\description{
Use Simpson's 3/8 rule to integrate a function
}
\details{
The \code{simp38} function uses Simpson's 3/8 rule to calculate the
integral of the function \code{f} over the interval from \code{a}
to \code{b}.  The parameter \code{m} sets the number of intervals
to use when evaluating.  Additional options are passed to the
function \code{f} when evaluating.
}
\examples{
f <- function(x) { sin(x)^2 + log(x) }
simp38(f, 1, 10, m = 10)
simp38(f, 1, 10, m = 100)
simp38(f, 1, 10, m = 1000)
}
\seealso{
Other integration: \code{\link{adaptint}};
  \code{\link{discmethod}}, \code{\link{revolution-solid}},
  \code{\link{shellmethod}}; \code{\link{gauss.hermite}},
  \code{\link{gauss.laguerre}},
  \code{\link{gauss.legendre}}, \code{\link{gaussint}};
  \code{\link{giniquintile}}; \code{\link{mcint}},
  \code{\link{mcint2}}; \code{\link{midpt}};
  \code{\link{romberg}}; \code{\link{simp}};
  \code{\link{trap}}

Other newton-cotes: \code{\link{adaptint}};
  \code{\link{giniquintile}}; \code{\link{midpt}};
  \code{\link{romberg}}; \code{\link{simp}};
  \code{\link{trap}}
}

