% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tridiagmatrix.R
\name{tridiagmatrix}
\alias{tridiagmatrix}
\title{Solve a tridiagonal matrix}
\usage{
tridiagmatrix(l, d, u, b)
}
\arguments{
\item{l}{vector of entries below the main diagonal}

\item{d}{vector of entries on the main diagonal}

\item{u}{vector of entries above the main diagonal}

\item{b}{vector of the right-hand side of the linear system}
}
\value{
the solution vector
}
\description{
use the tridiagonal matrix algorithm to solve a tridiagonal matrix
}
\details{
\code{tridiagmatrix} uses the tridiagonal matrix algorithm to solve a
tridiagonal matrix.
}
\seealso{
Other linear: \code{\link{cgmmatrix}},
  \code{\link{gaussseidel}}, \code{\link{iterativematrix}},
  \code{\link{jacobi}}; \code{\link{choleskymatrix}};
  \code{\link{detmatrix}}; \code{\link{invmatrix}};
  \code{\link{lumatrix}}; \code{\link{refmatrix}},
  \code{\link{rrefmatrix}}, \code{\link{solvematrix}};
  \code{\link{replacerow}}, \code{\link{rowops}},
  \code{\link{scalerow}}, \code{\link{swaprows}};
  \code{\link{vecnorm}}
}

