% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tridiagmatrix.R
\name{tridiagmatrix}
\alias{tridiagmatrix}
\title{Solve a tridiagonal matrix}
\usage{
tridiagmatrix(L, D, U, b)
}
\arguments{
\item{L}{vector of entries below the main diagonal}

\item{D}{vector of entries on the main diagonal}

\item{U}{vector of entries above the main diagonal}

\item{b}{vector of the right-hand side of the linear system}
}
\value{
the solution vector
}
\description{
use the tridiagonal matrix algorithm to solve a tridiagonal matrix
}
\details{
\code{tridiagmatrix} uses the tridiagonal matrix algorithm to solve a
tridiagonal matrix.
}
\seealso{
Other linear: \code{\link{choleskymatrix}},
  \code{\link{detmatrix}}, \code{\link{gdls}},
  \code{\link{invmatrix}}, \code{\link{iterativematrix}},
  \code{\link{lumatrix}}, \code{\link{refmatrix}},
  \code{\link{rowops}}, \code{\link{vecnorm}}
}
