% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vecnorm.R
\name{vecnorm}
\alias{vecnorm}
\title{Norm of a vector}
\usage{
vecnorm(b)
}
\arguments{
\item{b}{a vector}
}
\value{
the norm
}
\description{
Find the norm of a vector
}
\details{
Find the norm of a vector
}
\examples{
x <- c(1, 2, 3)
vecnorm(x)
}
\seealso{
Other linear: \code{\link{cgmmatrix}},
  \code{\link{gaussseidel}}, \code{\link{iterativematrix}},
  \code{\link{jacobi}}; \code{\link{choleskymatrix}};
  \code{\link{detmatrix}}; \code{\link{invmatrix}};
  \code{\link{lumatrix}}; \code{\link{refmatrix}},
  \code{\link{rrefmatrix}}, \code{\link{solvematrix}};
  \code{\link{replacerow}}, \code{\link{rowops}},
  \code{\link{scalerow}}, \code{\link{swaprows}};
  \code{\link{tridiagmatrix}}
}

