% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecnorm.R
\name{vecnorm}
\alias{vecnorm}
\title{Norm of a vector}
\usage{
vecnorm(b)
}
\arguments{
\item{b}{a vector}
}
\value{
the norm
}
\description{
Find the norm of a vector
}
\details{
Find the norm of a vector
}
\examples{
x <- c(1, 2, 3)
vecnorm(x)

}
\seealso{
Other linear: \code{\link{choleskymatrix}},
  \code{\link{detmatrix}}, \code{\link{gdls}},
  \code{\link{invmatrix}}, \code{\link{iterativematrix}},
  \code{\link{lumatrix}}, \code{\link{refmatrix}},
  \code{\link{rowops}}, \code{\link{tridiagmatrix}}
}
\concept{linear}
