% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wilkinson.R
\name{wilkinson}
\alias{wilkinson}
\title{Wilkinson's Polynomial}
\usage{
wilkinson(x, w = 20)
}
\arguments{
\item{x}{the \code{x}-value}

\item{w}{the number of terms in the polynomial}
}
\value{
the value of the function at \code{x}
}
\description{
Wilkinson's polynomial
}
\details{
Wilkinson's polynomail is a terrible joke played on numerical
analysis.  By tradition, the function is f(x) = (x - 1)(x - 2)...(x -
20), giving a function with real roots at each integer from 1 to 20.
This function is generalized and allows for \code{n} and the function
value is f(x) = (x - 1)(x - 2)...(x - n).  The default of \code{n} is
20.
}
\examples{
wilkinson(0)
}

